Refactors docs synchronization workflow
Refactors the documentation synchronization workflow to use a matrix strategy for deploying to both the `main` and `gh-pages` branches.

This change simplifies the workflow and reduces redundancy. It also updates the checkout action to v4 and upgrades the Trunk version.
Additionally, it adds logging using `loguru` and updates the pyproject.toml file, and modifies the resume file location and increments the view counter.

Updates README, icons, splash screen.